+++ src/user_pid.c
@@ -16,7 +16,7 @@ void mpid_init(mpid_c_t * plant_pid, int16_t max_out, int16_t min_out)
 	plant_pid->i_state = 0.0;
 }
 
-float mpid_update(mpid_c_t * plant_pid, int16_t plant_error, int16_t plant_current_value)
+float mpid_update(mpid_c_t * plant_pid, float plant_error, int16_t plant_current_value)
 {
 	float p_term;
 	float i_term;
